Introduce the world to the first teams in GSP
We've chosen Verify as our first team as it brings a semi complex setup
of the infrastructure.

We'd like to move it to Kubernetes and provide the team with concourse,
to help them run their pipelines.

Also, the main team would also be convenient to be manageable with the
resource.