Create a team setup pipeline
We'd like to have the concourse perform self-maintenace where possible.
Particulary when it comes to teams management.

The pipeline will read both, Concourse API and the `teams.yaml` file to
establish what teams are new and what teams have been deleted and adjust
itself to the new schema.

The concourse-teams-resource expects a specific format to be passed into
it. We'd like to prepare that file, with something that is fairly easily
managable - like a go application.