Spin up a cluster with Typhoon
We'd like to have a nice way of deploying clusters as and when we need
them.

We believe, Typhoon would be our preferred way of deploying clusters for
people including ourselves. It would give us nice room to improve and
modify as and when required.

This means we need a good way of defining new clusters therefore
providing a script to generate the template.

Documenting the use/approach to be repeatable by humans and perhaps be
adaptable to the CI/CD way.