Update custom module
We'd like to be able to define concourse username as well as password
for the generator.

We're now targeting the WIP branch of our for of typhoon, to enable EBS
volumes for kubernetes.

We've missed the user to team assignment in our values file, which
wouldn't be a problem if we didn't change the username in the first
place.