Use Terraform inside Docker to deploy
The primary driver behind putting Terraform in a container is to work
around a limitation of the `template` provider. The `template_dir`
resource generates files on the local file system and saves their
(absolute) path in the statefile.  This makes the project non-portable
between machines. By mounting the Terraform project directory at a
consistent mountpoint in the container the paths are always the same.