Use kube-system namespace
Most of the tools we're planning on delivering to all of the clusters,
will require some sort of super powers assigned. This is usually solved
by the `kube-system` namespace and in fact is hardcoded in some places.

This only leaves us with concourse which is the odd one out... We don't
feel confident nor excited about its placement.