allow declaring valueFileSecrets from flux module
The Helm Operator supports reading in values from a kubernetes
Secret[1], which is useful to us as it means we can "seal" secret
values and store them as part of the deployment chart.

This change allows declaring the name of the secret to read values from
when writing flux module stanzas.

[1] https://github.com/weaveworks/flux/blob/master/site/helm-integration.md#specvaluefilesecrets