add default values for chart_path and chart_ref
if your git repo places your chart at the root of the directory then the
chart_path must be set to empty-string. This makes this this default.

all our repos have a master branch and so we add a default for chart_ref
to be master to make it easier to write examples