Add Build & Run CI cluster
We'd like to build some docker images and perform set of tests in our
team.

We can spin up a cluster solely dedicated to that, with a cheaper and
`t3` single EC2 instance worker.

We'd like to make sure all of our stuff will also be deployed onto it.
Ingress for accessibility, prometheus and canary for monitoring, secrets
for pipelines.