generate initial valueFileSecret kubernetes Secret
when referencing a valueFileSecret that the helm-op will use to render
the chart, we need the secret to exist or it will not be able to apply.

this adds an empty secret for each valueFileSecret referenced.

I think this may be a problem in the future as it will mean that
applying the "addons" dir may not be idempotent. The solution may we to
seperate the deployment of secret-values from the deployment of the
charts.