require cluster.name and cluster.domain helm vars
We want to be able to use some per-cluster values in deployment charts.

An example of this is being able to know the domain that is mapped to
the cluster from the helm chart.

We have added two required values to the flux module which can be
accessible in templates via:

{{ .Values.cluster.name }} = the name given to the cluster
{{ .Values.cluster.domain }} = the fqdn mapped to the cluster's NLB