Pass in cluster_id as cluster_domain in canary mod
We recently added a requirement to pass in cluster_name and
cluster_domain to all HelmReleases. The cluster_id already passed in is
the same as the cluster_domain required, but we do not have the
cluster_name.

The canary chart doesn't care about these values at the moment so it's
safe to set to empty string