Make `sed` work with the BSD-userland `sed`
We're going to stop executing Terraform inside a container (and bringing
along a bunch of additional guff at the same time). We need to make this
script work with macOS `sed` rather than GNU `sed`. We might need to
make it work with both if we have people who join the team and don't use
macOS.