Stop using Terraform inside Docker
We're going to stop executing Terraform inside a container. The reasons
for doing this are documented in `731f551bd50785ad7805f2afbfc5c387147b6f87`.

This means the `initialise_canary_helm_codecommit.sh` script will
_probably_ run on macOS.  We therefore need to make this script work
with BSD-userland `sed` rather than GNU-userland `sed`.  We might need
to make it work with both if we have people who join the team and don't
use macOS.