Staging EIDAS cluster to watch staging of verify-eidas-deployment
As part of promotion, staging cluster should deploy from staging
branch of verify-edias-deployment.

We will want to do the same for production cluster to watch
production ranch of verify-eidas-deployment. We will do this when we
are ready to go live.

We output the github deployment public key from terraform which
is to be manually copied and added to the verify-eidas-deployment
with read and write access.