Add AWS HSM to EIDAS staging and production clusters
I made two design decisions:

- I only attempt to spin up a single HSM in the terraform code.
I could have set the count of the resource to be the number of
private subnets and the terraform would have likely failed the first
time, required the manual activation of the first HSM and then may
pass on a second run. However I discounted this option as I saw
strange behaviour in Terraform where it would create two HSMs but
would error on the third. I don't know if these two HSMs were
clones of each other or not. I was under the impression that you
should only be able to create a single HSM until you had activated
it so I wanted to avoid this confusing behaviour. Instead we
can either put in a PR later to bump the count of HSMs or scale
them manually using the UI.

- I chose to put the HSM creation code in this repo and not
in terraform-ignition. The HSM seemed very tied to EIDAS rather than
all clusters and that's why I lean towards keeping it in this
repository. On top of this, I looked into whether there were issues
with the HSM security group being created outside the gsp-cluster
and was able to mitigate any issues I came across. For example,
adding the HSM security group to our worker nodes can't be done
after the creation of the launch configuration for the worker nodes
as there is no Terraform resource for this. However, instead you
can add security group rules to the worker nodes security group
to allow them to talk to the HSM rather than needing to attach
the HSM security group to the instances themselves.