Revert to a state without KIAM
Sadly, the KIAM changes have somewhat broken down the Verify cluster.
For instance, the tool box is completely not accessible.

This is most likely with a side change enforcing the Canary to be always
enabled.

The clusters need to be running for the ongoing pen-tests and are not
required to have a canary just yet.

This should be then taken onto it's own story to establish the best ways
to proceed.