Add the role arn to assume when initialising the codecommit repo.
In the context of concourse, or when deferring the assume-role to
terraform as opposed to doing it outside like through aws-vault, the init
codecommit script doesn't inherit the assumed role environment so it would
fail to connect to codecommit with a 403. Pass the role arn down to the
updated gsp-cluster module so it can explicitly assume the right role.