Define empty provider
Due to some changes in the modules, we are no longer using the provider.
The existing terraform state however didn't have a chance to catch-up,
meaning it would fail to apply any changes.

Defining the empty provider and removing it after application has worked
in the past and we're following the known pattern.