Split up persistent terraform resources into an individual project
Design decisions:
- Needed to move the VPC resources to the persistent state
as the HSM relies on these to be created first before it can
be
- Added a new accounts/run-sandbox directory for our persistent
projects for dev stacks. This mimics the setup for eidas and in time
we should move our dev cluster projects to follow this pattern but
I kept that move out of scope for the moment to keep the PR small
- Didn't move the HSM module out of gsp-teams and into gsp-terraform-ignition
to again keep this PR as small as possible. Could be a future possible
move.
- Might not be happy with the naming of the modules e.g. `persistent`
but it might be good enough for the moment

Pair: @poveyd, @idavidmcdonald