Attach rules to existing security group
There is a default security group that is being created as part of the
HSM. Instead of creating a new one - which would likely tear down the
HSM cluster - we can create a new set of rules and attach them to that
magical existing group.

This is required for the workers to be able to interact with the HSM
Cluster.