Allow hsm access from within VPC subnets
This allows ingress to the HSM from anywhere inside the VPC private subnets.

Ideally we would only allow ingress from the "workers" security group to
lock it down a bit more, but this is not possible right now due to
ongoing work to split the hsm out into a seperate stage that is deployed
BEFORE the cluster exists.