Send CloudHSM logs to Cyber's Splunk
Applying this may take at least two runs if applying from scratch. The
Lambda log forwarder requires that the CloudWatch log group for the HSM
exists. Given this is created by the CloudHSM it's likely Terraform will
experience a race condition as this log group won't have been created.

To avoid this we could create the log group with the HSM module, and
`terraform import` the currently deployed state. We're going to not
bother doing this for the time being because the HSM creation is so
infrequent.