Map users and their GPG keys
We'd like not to rely on the GPG key servers which have a tendency to be
down very often. We'll be checking the git commit signatures against the
public keys stored in these yaml files at some point in the near future.

This will also be our definitive list of keys that are in use and
tracking of who has obtained one from RE.