Fix persistent remote state configuration.
Terraform silently swallows the HTTP 404 that gets returned for an
incorrectly configured state bucket, allowing the apply to continue. Only
for it to fail later with missing required parameter values.

The pipeline sets config on the terraform tasks which adds a workspace to
the remote state. This results in the clusters not using the correct url
to load the remote state. Setting the workspace should fix it, as per
the bootstrapper terraform.