Pass on the public keys to the verify pipelines
We'd like to automatically read in our `users/` directory and obtain the
public key for people who are allowed to sign the commits when
contributing to app repositories.

It needs to be managed through terraform that will compose a new file in
`addons/` directory and then apply it to the cluster.

These keys then will be loaded by our tenant's concourse and embedded
into their pipelines where needed to avoid unsigned commits being rolled
into production environment or being built.