Add the default to the terraform
The current state is only flowless if and when we're using a concourse
to deploy our clusters. Sadly, the dream of deploying our sandbox
cluster the same way is far away.

The easiest option that doesn't require poeple to be re-trained, is to
set a default for that variable.

The value, is simply an empty JSON array `[]` base64 encoded, to prevent
terraform's confusion with it's own implemetation of lists which happens
to be exactly the same... _cry_