populate promotion gpg keys from credentials
flux-releases can now be given a set of GPG public keys to verify
commits before deploying helm charts. this populates the
verification_keys with the values stored in concourse credential manager
(current ssm)

we also need to pass in the private key to the tools cluster so that
promotion tasks can sign the commits that will later be trusted by the
HelmRelease.