rename the commit-verification-keys var
there were two vars with the same name but different types (yaml vs
base64 string).

I renamed the var which is the list of all trusted developer gpg keys to
trusted-developer-keys which I feel is more explainitory.

I have removed the munging of the _other_ var entirely since yq does not
appear in the task-toolbox image so this has been failing