allows setting cluster-domain from cluster config
with cluster names being the same as account names we get weird
generated cluster-domains like portfolio.portfolio.govsvc.uk.

to avoid this while not being too prescriptive about exactly what the
domains are we are allowing setting the cluster domain manually from the
cluster config.

this sets the cluster domain/zone to {region}.{cluster_name}.govsvc.uk
and allows changing it in the future.