Scale down original worker ASG to 0
This scales down our original worker ASG to 0 nodes.

The `extra-workers-per-az-count` value guarantees that we will still
have at least 1 worker per AZ.

alphagov/gsp#600 added the cluster autoscaler, which I'm hoping will
scale up the per-AZ nodegroups to compensate for the loss of these
nodes.  Furthermore, it will scale the cluster to the right size for
the number & size of pods it's running.